# walking-safely/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Walking Safely API",
        "description": "API para o aplicativo Walking Safely - Navegação segura com análise de risco em tempo real",
        "contact": {
            "name": "Walking Safely Team",
            "email": "suporte@walkingsafely.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor de Desenvolvimento"
        },
        {
            "url": "https://api.walkingsafely.com/api",
            "description": "Servidor de Produção"
        }
    ],
    "paths": {
...
        "/timeseries": {
            "get": {
                "tags": [
                    "TimeSeries"
                ],
                "summary": "Série temporal de ocorrências",
                "description": "Retorna dados de série temporal para visualização em gráfico de linha",
                "operationId": "getTimeSeries",
                "parameters": [
                    {
                        "name": "region_id",
                        "in": "query",
                        "description": "Filtrar por região",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "region_ids",
                        "in": "query",
                        "description": "Filtrar por múltiplas regiões",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    },
                    {
                        "name": "crime_type_id",
                        "in": "query",
                        "description": "Filtrar por tipo de crime",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "crime_category_id",
                        "in": "query",
                        "description": "Filtrar por categoria de crime",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Data inicial",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "Data final",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "days",
                        "in": "query",
                        "description": "Últimos N dias",
                        "schema": {
                            "type": "integer",
                            "maximum": 365,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "granularity",
                        "in": "query",
                        "description": "Granularidade temporal",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "hour",
                                "day",
                                "week",
                                "month"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados da série temporal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/TimeSeriesPoint"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/timeseries/hourly": {
            "get": {
                "tags": [
                    "TimeSeries"
                ],
                "summary": "Padrão por hora do dia",
                "description": "Retorna análise de concentração de ocorrências por hora do dia",
                "operationId": "getHourlyPattern",
                "parameters": [
                    {
                        "name": "region_id",
                        "in": "query",
                        "description": "Filtrar por região",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "crime_type_id",
                        "in": "query",
                        "description": "Filtrar por tipo de crime",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Data inicial",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "Data final",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "days",
                        "in": "query",
                        "description": "Últimos N dias",
                        "schema": {
                            "type": "integer",
                            "maximum": 365,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Padrão horário",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "hour": {
                                                        "type": "integer",
                                                        "example": 14
                                                    },
                                                    "count": {
                                                        "type": "integer",
                                                        "example": 45
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/timeseries/daily": {
            "get": {
                "tags": [
                    "TimeSeries"
                ],
                "summary": "Padrão por dia da semana",
                "description": "Retorna análise de concentração de ocorrências por dia da semana",
                "operationId": "getDayOfWeekPattern",
                "parameters": [
                    {
                        "name": "region_id",
                        "in": "query",
                        "description": "Filtrar por região",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "crime_type_id",
                        "in": "query",
                        "description": "Filtrar por tipo de crime",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Data inicial",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "Data final",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "days",
                        "in": "query",
                        "description": "Últimos N dias",
                        "schema": {
                            "type": "integer",
                            "maximum": 365,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Padrão por dia da semana",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "day": {
                                                        "description": "0=Domingo, 6=Sábado",
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "day_name": {
                                                        "type": "string",
                                                        "example": "Segunda"
                                                    },
                                                    "count": {
                                                        "type": "integer",
                                                        "example": 120
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/timeseries/heatmap": {
            "get": {
                "tags": [
                    "TimeSeries"
                ],
                "summary": "Mapa de calor hora x dia",
                "description": "Retorna matriz de concentração combinando hora do dia e dia da semana",
                "operationId": "getHourDayHeatmap",
                "parameters": [
                    {
                        "name": "region_id",
                        "in": "query",
                        "description": "Filtrar por região",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "crime_type_id",
                        "in": "query",
                        "description": "Filtrar por tipo de crime",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Data inicial",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "Data final",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "days",
                        "in": "query",
                        "description": "Últimos N dias",
                        "schema": {
                            "type": "integer",
                            "maximum": 365,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Matriz hora x dia",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "hour": {
                                                        "type": "integer",
                                                        "example": 14
                                                    },
                                                    "day": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "count": {
                                                        "type": "integer",
                                                        "example": 25
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "TimeSeriesPoint": {
                "properties": {
                    "period": {
                        "type": "string",
                        "example": "2025-01-05"
                    },
                    "count": {
                        "type": "integer",
                        "example": 45
                    },
                    "cumulative": {
                        "type": "integer",
                        "example": 1250
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter your Bearer token in the format: Bearer {token}"
            }
        }
    },
...
}